requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncer==0.0.8
attrs==25.4.0
brotli==1.2.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cloudpickle==3.1.2
colorlog==6.10.1
cryptography==46.0.5
diskcache==5.6.3
distro==1.9.0
dspy==3.1.3
fastapi==0.135.1
fastuuid==0.14.0
ffmpy==1.0.0
filelock==3.25.1
frozenlist==1.8.0
fsspec==2026.2.0
gepa==0.0.26
google-auth==2.49.0
google-genai==1.66.0
gradio==6.9.0
gradio_client==2.3.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.13.0
json_repair==0.58.5
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
litellm==1.82.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
numpy==2.4.3
openai==2.26.0
optuna==4.7.0
orjson==3.11.7
packaging==26.0
pandas==3.0.1
pillow==12.1.1
propcache==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
safehttpx==0.1.7
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
stamina==25.2.0
starlette==0.52.1
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
tomlkit==0.13.3
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
